mkt_news_scrapper/requirements.txt
selenium==4.24.0
webdriver-manager==4.0.2
# mktnews ingestion
python-dotenv==1.1.1
supabase==2.6.0
openai==0.28.1
anthropic==0.40.0
